split html parse function to a separate import
