+++ lib/html.ts
@@ -1,0 +1,3 @@
+import sanitizeHtml from 'sanitize-html';
+
+export const stripHTML = htmlContent => sanitizeHtml(htmlContent, { allowedTags: [], allowedAttributes: {} });
